Interpolate coarse frc file tested & working
The subroutine 'ncdf_read_coarse_grid' to interpolate the 2D surface forcing from a frc file with half the resolution in x & y is now tested and confirmed to be working.

Minor differences (approx +/-5% near sharp gradients) is deemed to be due to the fact that the interpolation process is slightly different between interpolating forcing files offline and doing it in roms. For the offline forcing file the interpolation is done for lat/lon first and then the force itself. This creates additional errors and hence the difference in files.

The same interpolation algorithm was done directly in matlab to the full domain (and hence without the need to account for MPI decomposition), and the minor differences were still present, hence not a bug in the roms subroutine.

Removed debug flags ddevin_debug & ddevin_debug_frc from a few files as routine has now been confirmed as working.