makedepf90 fixed .f to .o only issue
Noticed that if you 'touch cppdefs.h' then the only thing that compiles is .f to .o, which is wrong because cppdefs.h will in turn change the .F file, so your compiled code is wrong.

Discussed with JM and solution is to delete all .f files forcing the depedent files to be recompiled from .F.
We decided it is cleaner to not have .f files floating around.
If you want to look at the raw code after c preprocessing (sometimes useful), you can still run mpc command and create .f file should you wish.