Code checking script implemented & working
Examples/code_check:
Script to check that the code still gives the same benchmark results in terms of the diagnostics variables.
-> Sums up all diagnostic terms and compares them to log file of perviously run simulation.
-> Bash script to run simulations
-> Python script for result checking by reading in .log files of roms terminal outputs.

Main use of this will be for working code that you want to reorganise or neaten up, but want to ensure nothing has been broken in the process.

This script should be run before anyone commits code!

Makedefs.inc:
- Added MPIHOME to VPATH & CPATH so that NHMG libraries work. Daniel discovered this issue.