Ncdf wrappers used on primary + aux variables
Used ncdf wrapper nc_define_var and nc_write_var on primary and auxiliary variable in def_his, wrt_his and wrt_avg. This needed to be done, but is also helpful as only need to change one line of code to make all outputs compressed.

Tried to use ncdf_create_file in def_his but had issues as the record counting is different from that of surf_flux output, so in the end kept seperate, just updated what was in def_his to use nf90 calls.

Have not 100% confirmed all variables are correct, will do that for next commit.
Also need to neaten up, and remove ncvars.h related stuff that is not needed anymore.

Makedefs.inc:
- Added compiler flags for when using vtune to get source code location and line numbers.