maya trailing spaces on vars & bgc vname issues
Wrappers work fine with variable name in 'vname' on comet and my laptop but not on maya since it is an old version of ifort. Need to stick with old lvar=lenstr(vname(1,indxZ)) method of old.

Tested new wrappers on Pierre's bgc pacific model and realized Akv's name was being overwritten by init_scalars_bec2. vname(4,indxPO4) this is only valid for ETH not ours because vname has only 3 rows in roms not 4 like ETH code.