ncjoin_mpi ready for use
- master only does pre-amble (checking input files, creating joined file).
- takes in arguements np_x and np_y for processor split for ncjoin_mpi sub-domains.
- added help info.
- added readme information in code on how to use ncjoin_mpi.
- added run script for correct use on expanse. (Note it only works on expanse!)

ncjoin_mpi is now working relatively well. On a pacific 12km (pacbig) simulatiion (1840*960*100)
with partial files of 10x10. we got the following results using d1 compression with shuffle:
   Program:       cores    time (s)
ncjoin (serial)     1       597
ncjoin_mpi 4x3     12       135
ncjoin_mpi 5x5     25        73
ncjoin_mpi 7x7     49        53
ncjoin_mpi 10x5    50        48
ncjoin_mpi 10x10  100        37

As is evident the scaling is far from perfect, but it is pleasing that we are getting a 10+ times
speed up using 50 cores. We found that the reading scales well but the writing does not.
See comments in code for further suggestions on independent vs collective writing.

Going to post on netCDF or HDF5 forum to ask why we might not be scaling so well.