Improved readmes for njoin_mpi
Updated readmes for using ncjoin_mpi as going to merge to master for everyone to use.
Still having unsolved issue for L4 simulation. ncjoin_mpi works for 64 cores but not 96 or 128.
It has worked fine for other domains so it is useable for now.

Added handling of all non-partionable array types.

Improved guidance on job scripts for optimal usage.