Added ability to write boundaries only
Currently we are writing out boundary diagnostics into full-domain arrays, with zeros everywhere but the boundary.

Added an example of writing u_west_bry (u_wbry). This is better for memory but unfortunately ncjoin doesn't like it's dimensions and would need to be update to handle this new type of variable.

Might just be easier to add bry contribution to one of the other diagnostic terms that isn't using its boundary points... Will confirm with JM.