Put in variables to confirm avg budget closes
I added terms for the sum of averages over time, and the gross difference in dz*u/v over the entire simulation.
This makes it easy to see if:

  dz*u(t=end) - dz*u(t=0) = sum ( Md_avg )

Over 2 steps it matches exactly, not to do more testing though.

This will be useful for people to turn on for a short test for their CPP configuration to ensure that diagnostic
budgets do match perfectly. They can then turn it off for the long simulation to save costs...