check_diag logical if user wants to confirm diags
The logical check_diag is used to output the total change in variable over timestep, the sum of the
diagnostics, and the difference between the two to confirm they match (as they should).
Improved the short names for these 3 terms and added long names.

Made the arrays allocatable for this check, so if check_diag is false then no memory is allocated.