Fixed closed boundary u/v FlxU/V at boundary 0.
With all the boundaries closed, the initial file is read in and the boundary is non-zero for the first timestep,
since u2dbc/u3dbc are only called further down the routine. This means a conservation test cannot be done from
the start of the model, since there is some flux over the boundary for the first timestep. This is now fixed.

Left in OBC_CHECK flag for debug output to show fluxes are zero at boundary. Will remove for next commit.

set_global_definitions:
- setup flag if fully closed boundaries OBC_NONE