New partit compiling
Sasha's version of partit on maya of /home/alex/tools is better than the partit I got from an
old version of ROMS. However, had issues getting the new partit to compile. It seems it didn't
like the -ip optimization which caused it to hang.

Overcame this issue by over-writing FFLAGS = -O3 in Tools-Roms/Makefile