Fixed interp coarse forcing bug
There was always a bug in ncdf_read_coarse_grid - an out of bounds array access. However, this did not corrupt any memory before, so it still worked. However, now that forcing is done using 'ncvs' derived types, the memory structure is slightly different and the bug caused issues.

I have fixed it to account for the single point computational margin at the boundaries, compared with 2 point margin in the interior.