Param module from param.h
Decided I will first migrate everything as is to match the .h file name with the new module name.
Once the migration is completed and working properly, then we can adjust the naming conventions
of the modules.

Param module gives exactly the same results in the netcdf output as before. only tested sample domain
with bulk_frc.