join_bry_child tool - improved speed
This tool was painfully slow before.
This was because every node and every variable for every record was tested to see if it contained the child variable.

Now, this is done at pre-processing and then later we only loop through nodes that actually contain a child var,
and then loop over only the specific variables the node contains.

On the sample domain for 1000 records it went from 26s to 4s.
I expect much more gain on larger domain since for the sample domain 4/6 nodes contained child bry,
whereas on a large domain with many partitions this ratio will be much lower.