Fix MPI_wait for some libraries
There was an 'illegal' call of MPI_wait that the compiler didn't complain about when we relied on "include mpif.h", but now that we rely on "use mpi", the compiler does some compile time checks and complains (only for some mpi libraries, not all).

The MPI_wait call now matches the standard and should work for all MPI libraries.