old read_write cleanup
Fixed cycle_length bug.

Spoke with Jeroen and we decide not a trivial fix to get new set_frc_data to work in current format,
because it doesn't do exchanges nor masking.
This means it has to be done manually every timestep, which is inefficient.

Furthermore, coarse2fine forcing interpolation is baked into old set_frc_var_tile and too much work now
for me to move this into set_frc_data as it is based on GLOBAL_2D_array sizes, etc.

So interim solution is to cleanup all unused functions in old read_write.F, and at some point it will be
deprecated fully in future...old read_write cleanup

Fixed cycle_length bug.

Spoke with Jeroen and we decide not a trivial fix to get new set_frc_data to work in current format,
because it doesn't do exchanges nor masking.
This means it has to be done manually every timestep, which is inefficient.

Furthermore, coarse2fine forcing interpolation is baked into old set_frc_var_tile and too much work now
for me to move this into set_frc_data as it is based on GLOBAL_2D_array sizes, etc.

So interim solution is to cleanup all unused functions in old read_write.F, and at some point it will be
deprecated fully in future...