Work-around for issue with NHMG
See previous commit. While removing the intent(out) attribute for
dummy pointer arguments allows the code to compile, it breaks the
performance of the mg solver. Changing the attribute to intent(inout)
appears to resolve the problem. Unfortunately, I'm not sure why this is
the case. It might be a bug in the compiler.