Scott B's fix to exact_restart
Instead of computing DU/DV_avg1 at the start of the run,
we need to read them from the restart file. This fixes
the exact restart for basic runs. There may be some additional
changes needed, such as reading in the ub_tuning coefficients
when tuned orlanski is used, and possibly others.