Bug fix to read_inp.
The attribute that stored the list of forcing file names could overflow due
to ia long list of file names or long path names to those files. I've stripped
the path names and the node number from the file names in the attribute, but kept
the attribute. Also, it was only writting the the node 0 file. Now it writes the
attribute to all partial files