Change vertical structure of forcing in cdr_frc.F
This fixes a problem in cdr_frc.F, where the code would place the same
amount of total forcing in a column regardless of how cdr_vsc and cdr_dep are set. This
was leading to extreme forcing concentrations in shallow water, or wherever
much of the forcing would have occurred beneath the seafloor.