Added handler for very small cdr_hsc in CDR forcing module.
Previously it was possible to have no CDR release in the model when the horizontal
scale was set to be a really small nonzero value (significantly smaller than the gridscale), even though
we had added a handler to deal with the most extreme case of hsc=0 (in which case the
model would just find the nearest cell center). This commit unifies the logic for both
of these cases, so that the CDR module always calculates the distance from the
release location to the nearest cell center, and will place the release there whether
hsc=0 or hsc is so small that no "significant" release points are close by. If hsc is large
then the code will behave as before.