SPONGE module was leading to broken Laplacian diffusivity.
When SPONGE was defined, the Laplacian diffusivities that were set in
the .in file were overridden with the diffusivity for temperature. So
basically the only diffusivity that mattered in the "tracer_diff2" field
was the first one.  This commit fixes that, and also cleans up a little
flotsam in set_nudgcof.F.