Flexible buffer size around arrays.
We used to have a hard coded buffer size of 2 for all arrays. That is
sufficient for 3rd upwind advection and the 4th order pressure gradient.
In order to facilite the introduction of higher order algorithms, the
GLOBAL_2D_ARRAY definition is now set to 1-bf:nx+bf for the x-direction
and equivalent for the y-direction. The value of bf is set in dimensions.F
and will move to dimensions.opt in the future, at which point param.opt/param.F
will be retired.