The intent was to collate the lmd_mixing stuff
Now also included are the move to include T and S as 'core'
ocean variables in terms of output control. Tracers.F now
controls the tracers past T and S.
Additionally, I got rid of the vname stuff for the core variables
as we don't actually allow variations on the names of these.